feat(metering): add wasmer-middleware for metering
As we're on Wasmer 4.X we're stuck with the metering implementation for
the same version. Sadly this contains an exploitable error in the
implementation, which was fixed in Wasmer 5.X and up. To avoid having to
update Wasmer we simply took the metering implementation and added it to
our repository, with some minimal changes to make Rust happy.